feat: refactor dataview components and update imports; add kanban_move_card functionality
